test: move consume_note to the crate root and reuse it more widely
Promotes the shared `consume_note` helper from `scripts/mod.rs` to `tests/lib.rs`
so both the `auth` and `scripts` test trees can use it via `crate::consume_note`.
Removes the identical copy in `auth/network_account.rs` and replaces the inline
consume-and-commit sequence in `scripts/fee_manager.rs` with a call to it.
